Add new version of Posts Extension, Posts2
For release GitHub:Release posts, we only need to match with existing template based on "repo-name"-"tag-name".xx, the posts date is fetched from the Tag commit.

The Posts2 Extension support adding Posts in given subfolder to the year/month/day structure based on page.date and not only date parsed from page-name.

In Posts2:

2011-01-01-name.xx

is handled the same as

name.xx
  ---
  date: 2011-01-01
  ---