Add page_authors(page) method to Awestruct::Extension::Guide
Returns a Array of unique author.name's based on the Git commit history located at page.site.dir for the given page.
The Array is ordered by number of commits done by the authors.

page_authors can be used as a Helper via:

   helper Awestruct::Extension::Guide

or if using the Guide extension, it will be used to populate the page.authors/guide.authors objects.

(note: if page.authors exists, the two Arrays will be merged placing page.authors first)