Update README.textile
Hi,

I added section "Books" and a link to the Arquillian in Action written by Alex Soto Bueno and Jason Porter for Manning Publications. We believe this book is a great resource and all members of community will benefit from this information. If you find this inappropriate, please let me know where I can publish it.